Use minikube registry addon hack
The minikube registry addon provides an in-cluster registry.
Unfortunately it does not appear to be possible to access it via the
same host name inside and outside the cluster.  Inside the cluster,
you can access by its service name, registry.kube-system.  Outside the
cluster, it is known as localhost:5000, provided by the
kube-registry-proxy.  This makes docker pushing and pulling
asymmetric, you push to registry/blah and pull from
localhost:5000/blah, which does not work.  Work around this by adding
an entry for registry.kube-system.svc.cluster.local in the minikube
VM /etc/host file with the IP address for the service.